+++ bin/fwup
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../lib/fwup.js');
