+++ lib/handle-exit.js
@@ -1,0 +1,6 @@
+module.exports = function handleExit(port) {
+  if(port.readable) {
+    port.close();
+  }
+  process.exit();
+}
